win_injector: seperate createprocess method
